Update README.md
the titanic EDA was about to get the insights from the data concerning the people who were involved in that titanic ship male, females and children. I wanted to review the number of deaths and also who survived the horror accident using the age, sex and other variables which were given.